+++ src/main/webapp/WEB-ROOT/js/jobApplication.js
@@ -59,7 +59,7 @@
                     "onBoardedTime",
                     "resignedTime",
                     "lastUpdatedTime",
-                    "<button  class='btn btn-info btn-sm editJobA'  ID='editJobA' onclick='updF(id)'><span class='glyphicon glyphicon-user'></span> 编辑</button> <button  class='btn btn-info btn-sm delJobA' ID='delJobA' onclick='delF(id)'><span class='glyphicon glyphicon-user'></span>删除</button>" ] ,
+                    "<button  class='btn btn-info btn-sm editJobA'  ID='editJobA' onclick='updF(id)'><span class='glyphicon glyphicon-pencil'></span> 编辑</button> <button  class='btn btn-info btn-sm delJobA' ID='delJobA' onclick='delF(id)'><span class='glyphicon glyphicon-remove'></span>删除</button>" ] ,
             name : ["ID", "姓名", "电话", "身份证号", "推荐人", "推荐人电话", "应聘企业", "应聘城市", "应聘职位", "求职/推荐时间", "最新状态", "入职时间", "离职时间", "处理时间", "_opt" ],
             tid : "id",
             checkBox : "id"
@@ -112,28 +112,31 @@
     }
 
     function delF(id) {
-        $("#alert").load("WEB-ROOT/html/common/alert.jsp",
-            function(response, status, xhr) {
-                $('#alert').html(response);
+        if (confirm("确定删除该记录？") == true) {
+            $("#alert").load("WEB-ROOT/html/common/alert.jsp",
+                function(response, status, xhr) {
+                    $('#alert').html(response);
+                });
+            $.ajax({
+                type : "DELETE",
+                url : "jobApplication?id="+id,
+                async : true,
+                contentType : "application/json;charset=utf-8",
+                success : function() {
+                    $("#alertText").text("删除成功");
+                    $("#alertModel").modal({
+                        keyboard : true
+                    });
+                    queryPage(1);
+                },
+                error : function(msg) {
+                    $("#alertText").text("删除失败");
+                    $("#alertModel").modal({
+                        keyboard : true
+                    });
+                }
             });
-        $.ajax({
-            type : "DELETE",
-            url : "jobApplication?id="+id,
-            async : true,
-            contentType : "application/json;charset=utf-8",
-            success : function(msg) {
-                $("#alertText").text(msg);
-                $("#alertModel").modal({
-                    keyboard : true
-                });
-            },
-            error : function(msg) {
-                $("#alertText").text("删除失败");
-                $("#alertModel").modal({
-                    keyboard : true
-                });
-            }
-        });
+        }
     }
 
     $("#addJobApplication").click(
@@ -216,7 +219,7 @@
             data: JSON.stringify(data),
             async: true,
             contentType: 'application/json;charset=utf-8',
-            success : function(msg) {
+            success : function() {
                  $( "#addAlert" ).load( "WEB-ROOT/html/common/alert.jsp", function( response, status, xhr ) {
                     $('#addAlert').html(response);
                     $("#alertText").text("操作成功");
@@ -224,6 +227,7 @@
                         keyboard: true
                     });
                  });
+                 queryPage(1);
             },
             error : function(msg) {
                 $( "#addAlert" ).load( "WEB-ROOT/html/common/alert.jsp", function( response, status, xhr ) {
