+++ src/main/webapp/WEB-ROOT/js/user.js
@@ -41,7 +41,7 @@
                     "name",
                     "account",
                     "authority",
-                    "<button  class='btn btn-info btn-sm editUser'  ID='editUser' onclick='updF(id)'><span class='glyphicon glyphicon-user'></span> 编辑</button> <button  class='btn btn-info btn-sm delUser' ID='delUser' onclick='delF(id)'><span class='glyphicon glyphicon-user'></span>删除</button>" ] ,
+                    "<button  class='btn btn-info btn-sm editUser'  ID='editUser' onclick='updF(id)'><span class='glyphicon glyphicon-pencil'></span> 编辑</button> <button  class='btn btn-info btn-sm delUser' ID='delUser' onclick='delF(id)'><span class='glyphicon glyphicon-remove'></span>删除</button>" ] ,
             name : ["ID", "用户姓名", "账号", "是否为管理员", "_opt" ],
             tid : "id",
             checkBox : "id"
@@ -88,28 +88,31 @@
     }
 
     function delF(id) {
-        $("#alert").load("WEB-ROOT/html/common/alert.jsp",
-            function(response, status, xhr) {
-                $('#alert').html(response);
+        if (confirm("确定删除该记录？") == true) {
+            $("#alert").load("WEB-ROOT/html/common/alert.jsp",
+                function(response, status, xhr) {
+                    $('#alert').html(response);
+                });
+            $.ajax({
+                type : "DELETE",
+                url : "user?id="+id,
+                async : true,
+                contentType : "application/json;charset=utf-8",
+                success : function() {
+                    $("#alertText").text("删除成功");
+                    $("#alertModel").modal({
+                        keyboard : true
+                    });
+                    queryPage(1);
+                },
+                error : function(msg) {
+                    $("#alertText").text("删除失败");
+                    $("#alertModel").modal({
+                        keyboard : true
+                    });
+                }
             });
-        $.ajax({
-            type : "DELETE",
-            url : "user?id="+id,
-            async : true,
-            contentType : "application/json;charset=utf-8",
-            success : function(msg) {
-                $("#alertText").text(msg);
-                $("#alertModel").modal({
-                    keyboard : true
-                });
-            },
-            error : function(msg) {
-                $("#alertText").text("删除失败");
-                $("#alertModel").modal({
-                    keyboard : true
-                });
-            }
-        });
+        }
     }
 
     $("#addUser").click(
@@ -174,7 +177,7 @@
             data: JSON.stringify(data),
             async: true,
             contentType: 'application/json;charset=utf-8',
-            success : function(msg) {
+            success : function() {
                  $( "#addAlert" ).load( "WEB-ROOT/html/common/alert.jsp", function( response, status, xhr ) {
                     $('#addAlert').html(response);
                     $("#alertText").text("操作成功");
@@ -182,6 +185,7 @@
                         keyboard: true
                     });
                  });
+                 queryPage(1);
             },
             error : function(msg) {
                 $( "#addAlert" ).load( "WEB-ROOT/html/common/alert.jsp", function( response, status, xhr ) {
